Helper functions for working with weeks (untested)
