+++ src/services/api-types.ts
@@ -15,21 +15,8 @@ export const CountAndProportionWithCISchema = zod.object({
   proportion: ValueWithCISchema,
 });
 
-const yearWeekRegex = /^(\d{4})-(\d{1,2})$/;
+export const yearWeekRegex = /^(\d{4})-(\d{1,2})$/;
 export const YearWeekSchema = zod.string().regex(yearWeekRegex);
-export function parseYearWeekString(
-  yearWeek: zod.infer<typeof YearWeekSchema>
-): { year: number; week: number } {
-  const m = yearWeek.match(yearWeekRegex);
-  if (!m) {
-    throw new Error('invalid YearWeek string');
-  }
-  const parsed = { year: +m[1], week: +m[2] };
-  if (!(parsed.week >= 1 && parsed.week <= 53)) {
-    throw new Error('invalid week in YearWeek string');
-  }
-  return parsed;
-}
 
 export const YearWeekWithDaySchema = zod.object({
   yearWeek: YearWeekSchema,
@@ -128,6 +115,7 @@ export interface Selection {
 // TypeScript types from schemas
 
 export type ValueWithCI = zod.infer<typeof ValueWithCISchema>;
+export type YearWeekWithDay = zod.infer<typeof YearWeekWithDaySchema>;
 export type Country = zod.infer<typeof CountrySchema>;
 export type Sample = zod.infer<typeof SampleSchema>;
 export type SampleResultList = zod.infer<typeof SampleResultListSchema>;
